+++ tests/standard_state_test.py
@@ -3,7 +3,7 @@ import sys
 import io
 import pytest
 
-from behavior_machine.board import Board
+from behavior_machine.core import Board
 from behavior_machine.core import State, StateStatus, Machine
 from behavior_machine.library import WaitState, IdleState, PrintState
